requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.2
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.7
django-webpack-loader==0.6.0
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
kombu==4.3.0
Markdown==2.6.3
MarkupSafe==1.1.1
musicbrainzngs==0.5
psycopg2==2.7.7
Pygments==2.1.3
pytz==2018.9
PyYAML==5.1
requests==2.21.0
selenium==3.141.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
whitenoise==4.1.2
graphene==2.0